Properly handle async errors
Express is not optimized for async/await, so async errors are currently
bubbled up to the main process and not properly handled. You can read
more about Express error handling here:
http://thecodebarbarian.com/80-20-guide-to-express-error-handling

Since Probot is so async/await heavy, I think it makes sense to pull in
this extension that properly handles rejected promises in Express.