Adds missing awaits
Both `get` and `set` are async in the metadata library. The `await` in `get` is required in this case because otherwise it will always to set a promise on the metadata, which doesn't make sense, and the await in set is required to not swallow the rejection if it fails.