change phrasing on how to get the private key
Having the private key download at the end leverages the image provided for 'APP_ID', in order to cue and instruct the user on how to get the `.PEM` file; and groups the steps to edit `.env` together, leaving the download for last. This enhances both the setup flow and its legibility.

Thought it would also be useful to add a link here to see the additional environmental configurations available.