fix(package): remove "prepare" step
as part of the automated release with semantic-release, we replace a placeholder version string in lib/index.js. We do not want to read the version from package.json because it is not supported by ES Modules natively.

Unfortunately the "prepare" step is run right before the publish step, and after the replacements took place. So the easiest ist to just remove it and figure out a better way to install probot from github