feat: fix broken tests and removes default value
I have not added any new tests, I just fixed the old ones.
During this process I noticed, that for some reason the listen()
behavior changes when 0.0.0.0 (the previous default value) was used.
I still dont know why this happens, but found another issue with
the default value: Missing Ipv6 support.
I removed the state properties and listen() is now not called with a second
argument when bindAddress is undefined.