fix: correctly import (transpiled) app function for run and receive
There had been a fix previously #1448, to load transpiled app functions
for `probot run`.
This commit replaces the custom check for `.default` and `.__esModule`
properties with a dynamic import, which TypeScript compiles to a
`require` call that is capable of handling both, commonJS and ESM, as
long as `esModuleInterop` is set to `true` in probot's root
`tsconfig.json`.
Since a dynamic import is async, I had to wrap the `probot receive` CLI
in an async function.